Adds `PanicOnError` as a recoverable version of `Must`. (#49)
Adds `PanicOnError` as a recoverable version of `Must`.
